TakePhoto/RecordingのUIのリファクタリング。
